+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Server.Host` returns the canonical host string for the given domain.
 - `Forbidden` and `ForbiddenWithError` replies with 403 Forbidden.
 - `WithSite` is a helper to set the site in the context. Primarily useful in tests.
+- `SetCanonicalLogField` adds a field to the canonical log line.
 
 ### Fixed